requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
python-jobspy>=1.1.0
supabase>=2.0.0
requests>=2.31.0
rapidfuzz>=3.0.0
